% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baptransform.R
\name{laplaceBAP}
\alias{laplaceBAP}
\title{Laplace transform for BAP}
\usage{
laplaceBAP(tt, mu, kappa, sigma, lambda = 0, zeta = 1, w = -1)
}
\arguments{
\item{tt}{Vector of times.}

\item{lambda}{Jump arrival rate of process.}

\item{zeta}{Mean of jump size distribution.}

\item{w=-1}{Auxiliary parameter in transform.}
}
\value{
List with (tt, A0, B0, A1, B1), all vectors of equal length.
}
\description{
Laplace transform of Y[t] and time-derivative of transform.
}

