% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{laplaceBAPmontecarlo}
\alias{laplaceBAPmontecarlo}
\title{Laplace transform for BAP process by simulation}
\usage{
laplaceBAPmontecarlo(X0, tt, mu, kappa, sigma, lambda = 0, zeta = 1,
  w = -1, trials = 10000)
}
\arguments{
\item{X0}{Initial condition.}

\item{tt}{Vector of times.}

\item{mu}{Drift of process.}

\item{kappa}{Mean reversion of process.}

\item{sigma}{Volatility of process.}

\item{lambda}{Jump arrival rate of process.}

\item{zeta}{Mean of jump size distribution.}

\item{w=-1}{Auxiliary parameter in transform.}

\item{trials=10000}{Number of trials in simulation}
}
\value{
psi vector of equal length to tt.
}
\description{
Laplace transform of Y[t]
}
