% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{laplaceCIRmontecarlo}
\alias{laplaceCIRmontecarlo}
\title{Laplace transform for CIR process by simulation}
\usage{
laplaceCIRmontecarlo(X0, tt, mu, kappa, sigma, w = -1, trials = 10000)
}
\arguments{
\item{X0}{Initial condition.}

\item{tt}{Vector of times.}

\item{mu}{Drift of process.}

\item{kappa}{Mean reversion of process.}

\item{sigma}{Volatility of process.}

\item{w=-1}{Auxiliary parameter in transform.}
}
\value{
psi vector of equal length to tt.
}
\description{
Laplace transform of Y[t]
}
