% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rBAPtransition}
\alias{rBAPtransition}
\title{Draw from BAP transition density, i.e., X[t] given X[0]}
\usage{
rBAPtransition(X0, mu, kappa, sigma, lambda = 0, zeta = 1, timestep)
}
\arguments{
\item{X0}{Initial conditions (scalar).}

\item{lambda}{Jump arrival rate of process.}

\item{zeta}{Mean of jump size distribution.}

\item{timestep}{time t}
}
\value{
Xt Value of processes at time t.
}
\description{
Draw from BAP transition density, i.e., X[t] given X[0]
}

