% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rBAPtransition2}
\alias{rBAPtransition2}
\title{Draw (X[t],Y[t]) jointly from BAP transition density.}
\usage{
rBAPtransition2(XY0, mu, kappa, sigma, lambda = 0, zeta = 1, horizon)
}
\arguments{
\item{XY0}{Initial condition list(X=X0,Y=Y0)}

\item{lambda}{Jump arrival rate of process.}

\item{zeta}{Mean of jump size distribution.}

\item{horizon}{time t}
}
\value{
List (X=Xt,Y=Yt) of process at time t.
}
\description{
Draw (X[t],Y[t]) jointly from BAP transition density.
}

