% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rBessel}
\alias{rBessel}
\title{Draw a Bessel random variable.}
\usage{
rBessel(nu, z)
}
\arguments{
\item{nu}{First parameter of distribution}

\item{z}{Second parameter of distribution}
}
\value{
BES(nu,z) random variate
}
\description{
Method on page 285 of G-K
}

