% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rCIRpath}
\alias{rCIRpath}
\title{Draw path from CIR transition density on [0,horizon]}
\usage{
rCIRpath(X0, mu, kappa, sigma, horizon, ticks)
}
\arguments{
\item{X0}{Initial condition.}

\item{mu}{Drift of process.}

\item{kappa}{Mean reversion of process.}

\item{sigma}{Volatility of process.}

\item{ticks}{Number of slices in [0,horizon]}

\item{tt}{horizon Final time.}
}
\value{
List of (tt,Xt), each of length ticks+1.
}
\description{
Evenly spaced times points 0,dt,...,horizon, where dt=horizon/ticks.
}
