% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rCIRstationary}
\alias{rCIRstationary}
\title{Draw from CIR stationary distribution.}
\usage{
rCIRstationary(mu, kappa, sigma, n = 1)
}
\arguments{
\item{mu}{Drift of process.}

\item{kappa}{Mean reversion of process.}

\item{sigma}{Volatility of process.}

\item{n=1}{number of observations.}
}
\value{
X Value of process.
}
\description{
If \code{n}>1, return will be a vector of length \code{n},
with vector elements draw independently. This is convenient
for simulation of many independent paths simulataneously.
}

