% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rCIRtransition}
\alias{rCIRtransition}
\title{Draw from CIR transition density}
\usage{
rCIRtransition(X0, mu, kappa, sigma, timestep)
}
\arguments{
\item{X0}{Initial condition.}

\item{mu}{Drift of process.}

\item{kappa}{Mean reversion of process.}

\item{sigma}{Volatility of process.}

\item{timestep}{time t}
}
\value{
Value Xt of process at time t.
}
\description{
Draw X[t] given X[0].
}
\details{
If \code{X0} is a vector, return \code{Xt} will be a vector of the
same length. The vector elements are draw independently. This allows
us to simulate many independent paths simulataneously.
}

