% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rCIRtransition2}
\alias{rCIRtransition2}
\title{Draw ([X[t],Y[t]) jointly from CIR transition density}
\usage{
rCIRtransition2(XY0, mu, kappa, sigma, horizon)
}
\arguments{
\item{XY0}{Initial condition list(X=X0,Y=Y0)}

\item{mu}{Drift of process.}

\item{kappa}{Mean reversion of process.}

\item{sigma}{Volatility of process.}

\item{horizon}{time t}
}
\value{
List (X=Xt,Y=Yt) of process at time t.
}
\description{
Draw (X[t],Y[t]) given (X[0],Y[0]).
}

