% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rMRCPtransition}
\alias{rMRCPtransition}
\title{Draw from MRCP transition density, i.e., X[t] given X[0]}
\usage{
rMRCPtransition(X0, mu, kappa, lambda = 0, zeta = 1, timestep)
}
\arguments{
\item{X0}{Initial conditions (scalar).}

\item{mu}{Drift of process.}

\item{kappa}{Mean reversion of process.}

\item{lambda}{Jump arrival rate of process.}

\item{zeta}{Mean of jump size distribution.}

\item{timestep}{time t}
}
\value{
Xlist list containing X[timestep], X[jumptimes-], jump times and sizes.

Xt Value of processes at time t.
}
\description{
Draw from MRCP transition density, i.e., X[t] given X[0]
}

