% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rPoissonTimes}
\alias{rPoissonTimes}
\title{Helper function to draw Poisson jump times out to horizon.}
\usage{
rPoissonTimes(lambda, horizon)
}
\arguments{
\item{lambda}{Jump arrival rate of process.}

\item{horizon}{time t}
}
\value{
Jump times
}
\description{
Final jump time will be beyond horizon.
}
