% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rBAPtransition}
\alias{rBAPtransition}
\title{Draw from BAP transition density, i.e., X[t] given X[0]}
\usage{
rBAPtransition(X0, mu, kappa, sigma, lambda = 0, zeta = 1, horizon)
}
\value{
Xt Value of processes at time t.
}
\description{
Draw from BAP transition density, i.e., X[t] given X[0]
}
