% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rCIRintegrated}
\alias{rCIRintegrated}
\title{Draw Y[t] conditional on (X[0],X[t]).}
\usage{
rCIRintegrated(X0, Xt, mu, kappa, sigma, horizon, K = 10)
}
\arguments{
\item{X0}{Initial condition.}

\item{Xt}{Final value at horizon.}

\item{mu}{Drift of process.}

\item{kappa}{Mean reversion of process.}

\item{sigma}{Volatility of process.}

\item{horizon}{time t}

\item{K}{number of terms in G-K expansion}
}
\value{
Value Yt of process at time t.
}
\description{
Glasserman and Kim FS 2011, Theorem 2.2
}

