% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapsim.R
\name{rCIRintegrated}
\alias{rCIRintegrated}
\title{Draw Y[t] conditional on (X[0],X[t]).}
\usage{
rCIRintegrated(X0, Xt, mu, kappa, sigma, horizon, GKterms = 10)
}
\arguments{
\item{X0}{Initial condition.}

\item{Xt}{Final value at horizon.}

\item{mu}{Drift of process.}

\item{kappa}{Mean reversion of process.}

\item{sigma}{Volatility of process.}

\item{horizon}{time t}

\item{GKterms}{number of terms in Glasserman-Kim expansion}
}
\value{
Value Yt of process at time t.
}
\description{
Algorithm is based on Theorem 2.2 in
\href{http://dx.doi.org/10.1007/s00780-009-0115-y}{Glasserman and Kim (2011)}.
}
